Swap unit & e2e test cases
